+++ packages/PlantimIcons/Sources/PlantimIcons/PlantimIcon.swift
@@ -2,8 +2,8 @@ import SwiftUI
 import Foundation
 
 public enum PlantimIconsMetadata {
-    public static let version = "2.0.1"
-    public static let registryHash = "dd50bd6712fa6917a5c8c9b3ba0ff65924ab44a12dc5812923238fa0eee2cbfb"
+    public static let version = "2.1.0"
+    public static let registryHash = "6e37c9e631574cc64b8172b3c63adafb794e18829adf58b7a72547fb4d0adbd7"
 }
 
 public enum PlantimIconName: String, CaseIterable, Sendable {